client_qt/requirements_qt.txt
anyio==4.9.0
certifi==2025.4.26
ecdsa==0.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.2.1
psutil==7.0.0
pyasn1==0.6.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-jose==3.5.0
qrcode==8.2
rsa==4.9.1
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
